move AA consistency tests to correct module
